minor fixes

+++ bin/cleanup.sh
@@ -15,35 +15,49 @@
 #-----------------------------------------------------------------------------
 
 #cd $WRK
-odir=$(grep 'OutputDIR  =' supermopex.py | cut -d\' -f2 | tr -d \/)
-rdir=$(grep 'RawDataDir =' supermopex.py | cut -d\' -f2 | tr -d \/)
+if [ -e supermopex.py ]; then
+	odir=$(grep 'OutputDIR  =' supermopex.py | cut -d\' -f2 | tr -d \/)
+	rdir=$(grep 'RawDataDir =' supermopex.py | cut -d\' -f2 | tr -d \/)
+else
+	echo "supermopex.py not found; "
+	odir=Products
+	rdir=Data
+fi
 
-#echo $rdir; echo $odir # ; exit
-
+# move or delete?
 if [ $# -eq 1 ]; then
 	if [ -d $1 ]; then
 		echo " move products to $1"
-		keep=1
+		keep="T"
 	else
 		echo " remove products ..."
-		keep=0
+		keep="F"
 	fi
 fi
 
+# check that raw data dir exists
+if [ ! -d $rdir ]; then
+	echo "ERROR: $rdir not found - quitting!"
+	exit 1
+fi
+
+# delete intermediate products from raw data dir
 for f in $rdir/r*/ch?/bcd; do 
 	echo "- now clean $f"
 	rm -f $f/SPI*_[f-z]*.fits $f/SPI*.tbl
 done
 
-echo " remove command etc. files"
-if [ $keep -eq 1 ]; then
+if [[ $keep == "T" ]]; then
+	echo "## move command etc. files to $1"
 	mv *.sh *.py *.out *.qall *.err addkey* FIF.tbl $1
-	mv medians $odir  cdf  supermopex.py  irac.log $1
+	mv medians $odir irac.log make_tiles* countFiles.dat  $1
 	rm -rf temp __pycache__ 
 else
-	rm -f *.sh *.py *.out *.qall *.err addkey* FIF.tbl cdf/log_*  # supermopex.py
-	rm -rf temp __pycache__ medians/* $odir/* countFiles.dat irac.log
+	echo "## delete command etc. files"
+	rm -f *.sh *.py *.out build.* *.err addkey* FIF.tbl cdf/log_*  # supermopex.py
+	rm -rf temp __pycache__ medians/* $odir/* countFiles.dat irac.log make_tiles*
 fi
+
 
 exit 0
 #rm *.lst *.fits *stars
